requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
alembic
cryptography
pyjwt